lecture management api done

+++ server/utilities/deleteFilesBypath.js
@@ -1,12 +1,13 @@
 const fs = require("fs");
 
 const deleteFileByPath = (filePath) => {
-  fs.unlink(filePath, (err) => {
-    if (err) {
-      throw err;
-    }
-  });
-  
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (err) => {
+      if (err) {
+        throw err;
+      }
+    });
+  }
 };
 
 module.exports = { deleteFileByPath };
